mtk-perf: fpsgo: prevent free but unfinished
The memory recycling is a must since memory leakage is possible
if we depends on timer/wq/deq callback.
However, when rescue is started but not yet finished,
the record should not be freed.
Enlarge the traversal period to prevent timer/work from unfinished
due to abnormal environment.

MTK-Commit-Id: 3afc46e9b11887ef5a527b48fb4671dbfbe96645

Change-Id: Iff2e11701719e76d71c603a44b344d1afcf9037b
Signed-off-by: ChiaYu Ku <ChiaYu.Ku@mediatek.com>
CR-Id: ALPS05038055
Feature: FPSGO